Fix Coverity: unsigned underflow, type mismatch, dead code, fread
Unsigned underflow: CGroupTable::DeleteGroup had unsigned pos
decremented past zero — use signed local (same fix as InsertGroup).

Type mismatch: CPropCalendarGeneral compared uint32_t size against
ULONG_MAX (always true on LP64, sentinel never matched). Changed
to UINT32_MAX. Also fixed CCalendarStoreNode initialization.

Dead assignments and missing else: CFileTable::AddPart had two
dead assignments to new_woRow. CRFC822Parser::ParseMultipart and
CAddressBook::ImportAddress assigned return values only needed
for side effects. CBodyTable::GetAttachIcon had missing else —
IsNotCached result was always overwritten by subsequent chain.
CTextDisplay removed dead unsigned < 0 checks and associated
dead else branches.

Unchecked fread/ftell: CAdbkIOPlugin and CVCardAdbkIOPlugin
DoImport did not check ftell for -1 error and ignored fread
return value. UNX_LFileStream::GetLength stored lseek result
in unsigned UInt32 — use off_t with error check.

+++ Sources_Common/Calendar_Store/CCalendarStoreNode.cpp
@@ -57,7 +57,7 @@ CCalendarStoreNode::CCalendarStoreNode()
 	SetFlags(eIsDirectory, true);
 	SetFlags(eHasExpanded, true);
 	mCalendarRef = iCal::cCalendarRef_Invalid;
-	mSize = ULONG_MAX;
+	mSize = UINT32_MAX;
 	mLastSync = 0;
 	mACLs = NULL;
 }
@@ -77,7 +77,7 @@ CCalendarStoreNode::CCalendarStoreNode(CCalendarProtocol* proto)
 	mName = proto->GetCalendarAccount()->GetBaseRURL();
 	mShortName = mName;
 	mDisplayName = proto->GetAccountName();
-	mSize = ULONG_MAX;
+	mSize = UINT32_MAX;
 	mLastSync = 0;
 	mACLs = NULL;
 }
@@ -95,7 +95,7 @@ CCalendarStoreNode::CCalendarStoreNode(CCalendarProtocol* proto, CCalendarStoreN
 	SetFlags(eIsOutbox, is_outbox);
 	mCalendarRef = iCal::cCalendarRef_Invalid;
 	mName = name;
-	mSize = ULONG_MAX;
+	mSize = UINT32_MAX;
 	mLastSync = 0;
 	mACLs = NULL;
